rename: ws/chat.ts -> socket/chat.ts
#35
